refactor: ajustes de nome e parametro dos endpoints
