+++ src/main/java/br/fiap/projeto/contexto/identificacao/adapter/controller/rest/exception/ClienteErrorHandler.java
@@ -16,23 +16,19 @@ public class ClienteErrorHandler {
 
     @ExceptionHandler({EntidadeNaoEncontradaException.class})
     public ResponseEntity<MensagemErroDTO> entidadeNaoEncontradaHandler(EntidadeNaoEncontradaException ex) {
-
         ResponseEntity<MensagemErroDTO> erro = ResponseEntity.status(HttpStatus.NOT_FOUND).body(new MensagemErroDTO(ex.getCode(), ex.getMessage()));
         return erro;
     }
 
     @ExceptionHandler({EntradaInvalidaException.class})
     public ResponseEntity<MensagemErroDTO> entradaInvalidaHandler(EntradaInvalidaException ex) {
-
         ResponseEntity<MensagemErroDTO> erro = ResponseEntity.status(HttpStatus.PRECONDITION_FAILED).body(new MensagemErroDTO(ex.getCode(), ex.getMessage()));
         return erro;
     }
 
     @ExceptionHandler({Exception.class})
     public ResponseEntity<MensagemErroDTO> genericHandler(EntradaInvalidaException ex) {
-
         ResponseEntity<MensagemErroDTO> erro = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new MensagemErroDTO(BaseException.DEFAULT_CODE, ex.getMessage()));
-
         log.error("Exceção não tratada: ", ex);
         return erro;
     }
