+++ src/main/java/br/fiap/projeto/contexto/identificacao/usecase/exception/BaseException.java
@@ -1,14 +1,11 @@
 package br.fiap.projeto.contexto.identificacao.usecase.exception;
 
 public abstract class BaseException extends Exception {
-
     public final static int DEFAULT_CODE = 4000;
-
-    private int code;
-    private String message;
+    private final int code;
+    private final String message;
 
     public BaseException(int code, String message) {
-
         this.code = code;
         this.message = message;
     }
